Moved the image to one of our servers
Was getting a console warning in Chrome Dev Tools that the certificate of the mailchimp site from which we were hosting the Save the Date image was going to be mistrusted in a future version of Chrome. Moved the image to one of our own web servers to remove the warning.